requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
defusedxml==0.7.1
distro==1.9.0
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.8
httpie==3.2.4
httpx==0.28.1
idna==3.10
jiter==0.9.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.4.3
openai==1.75.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
